Add example for setting default language (#268)
On the Multiple Languages support topic, added a paragraph and table to show the example.

If your form defines multiple languages and you want to set one language for data collection, set an explicit default language. This is done on the XLSForm's **settings** sheet by specifying the language in the **default_language** column. In the example [le below, the default language will be French.

On the media section, no changes made.